requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
coverage==7.13.5
cryptography==46.0.7
defusedxml==0.7.1
dj-database-url==3.1.2
dj-static==0.0.6
Django==6.0.4
django-allauth==65.15.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.33.1
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
static3==0.7.0
urllib3==2.6.3
